Change `trx_seq` type to 32 bit uint

+++ include/row/row.hpp
@@ -111,7 +111,7 @@ public:
 
    // Tables
     struct [[eosio::table("authorities")]] authority {
-        uint64_t             trx_seq   = 0; // sequence number of the last proposed transaction.
+        uint32_t             trx_seq   = 0; // sequence number of the last proposed transaction.
         uint32_t             threshold = 1;
         std::vector<authkey> keys;
         bool weights_cross_threshold(uint32_t weights) const { // does the weights reach threshold
@@ -123,7 +123,7 @@ public:
     struct [[eosio::table]] proposal {
         name                      proposal_name;
         time_point                create_time;
-        uint64_t                  trx_seq;    // transacion sequencial number i.e. authority.trx_seq += 1 at proposal
+        uint32_t                  trx_seq;    // transacion sequencial number i.e. authority.trx_seq += 1 at proposal
         std::vector<char>         packed_transaction;
         std::optional<time_point> earliest_exec_time;
         uint64_t primary_key() const { return proposal_name.value; }
